ComponentDcsId annotation created (detailed description below). DcsFact class created (detailed description below). And DcsAlertFact created - this one informs about any alerts coming up or going away or being acknowledged in DCS.
========= documentation for ComponentDcsId =========
Class (fact) properties that are annotated with this annotation are told to be carrying a component DCS ID (DCS datapoint element name), which is resolved by org.cern.cms.csc.dw.ws.ComponentPropertyResolver through ontology component DCSId data property to two things - component and component attribute. The search algorithm is looking for a component having a data property DCSId which *best* matches the provided DCS ID, the unmatched part of the ID is considered to be a component attribute. e.g. if you provide CSC_ME_M21_C01_HV.data.v1.status, it is matched to: component "ME-2/1/01/1/HVSegment1" and it's attribute "status".

======= documentation for DcsFact ========
 This class extends Fact class and adds componentDcsId (transient) and componentAttribute properties, so if you want to be able to refer to a component by it's DCS ID (i.e. DCS Datapoint Element), use this class as a base for your concrete fact class. When using this class as a base one can refer to components either by their ontologt ID using the component_id property or by DCS ID using the component_dcs_id property.